Stable version for GP

+++ core/network/src/main/java/com/aralhub/network/models/driver/NetworkActiveRideByDriverResponse.kt
@@ -1,7 +1,6 @@
 package com.aralhub.network.models.driver
 
 import com.aralhub.network.models.offer.Locations
-import com.aralhub.network.models.websocketclient.ClientRideResponseOptions
 import com.aralhub.network.models.websocketclient.ClientRideResponsePassenger
 import com.aralhub.network.models.websocketclient.ClientRideResponsePaymentMethod
 import com.google.gson.annotations.SerializedName
@@ -12,7 +11,7 @@ data class NetworkActiveRideByDriverResponse(
     val status: String,
     val amount: Int,
     @SerializedName("wait_amount")
-    val waitAmount: Double,
+    val waitAmount: Int,
     val distance: Float,
     val locations: Locations,
     @SerializedName("is_active")
